Modifying navigation logic and hierarchy
Was running into issues with previous config that every route was causing a full page reload which was clashing with the session logic to display logged in accounts through the materialui elements. New config uses ReactRouterAppProvider instead of MaterialUIs AppProvider which was found to be the cause of the refresh issue. Also now using Page Containers with an eye to use it in future styling needs